workerize parsing of time series data
