+++ lib/components/TimeSeriesViewer/TimeSeriesViewerContext.js
@@ -35,6 +35,8 @@ var _NeonEnvironment = _interopRequireDefault(require("../NeonEnvironment/NeonEn
 
 var _rxUtil = require("../../util/rxUtil");
 
+var _parseTimeSeriesData = _interopRequireDefault(require("../../workers/parseTimeSeriesData"));
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _getRequireWildcardCache() { if (typeof WeakMap !== "function") return null; var cache = new WeakMap(); _getRequireWildcardCache = function _getRequireWildcardCache() { return cache; }; return cache; }
@@ -151,29 +153,10 @@ var generateYAxisRange = function generateYAxisRange() {
   }
 
   return axisRange;
-}; // Functions to convert a value to the proper JS data type given a NEON variable dataType
-
-
-var castFloat = function castFloat(v) {
-  var cast = parseFloat(v, 10);
-  return Number.isNaN(cast) ? null : cast;
-};
-
-var castInt = function castInt(v) {
-  var cast = parseInt(v, 10);
-  return Number.isNaN(cast) ? null : cast;
-};
-
-var DATA_TYPE_SETTERS = {
-  dateTime: function dateTime(v) {
-    return typeof v === 'string' ? v.replace(/"/g, '') : v;
-  },
-  real: castFloat,
-  'signed integer': castInt,
-  'unsigned integer': castInt
 }; // PropTypes for any Tab Component (or component within a tab) for gettingsetSelectedTab
 // and TAB_IDS as props. This is used by anything in a tab wanting to afford the ability to
 // route the user to another tab.
+
 
 var TabComponentPropTypes = {
   setSelectedTab: _propTypes.default.func.isRequired,
@@ -706,99 +689,6 @@ var parseSitePositions = function parseSitePositions(site, csv) {
   return newSite;
 };
 /**
- * Build an object containing series of data, indexed by fieldName, from a data fetch response
- * We don't use the same CSV parse method as with variables/positions because we don't want rows:
- *   [ { varX: 1, varY: 2, varZ: 3 }, { varX: 10, varY: 20, varZ: 30 }, ... ]
- * ...we want columns to individually stitch together with other months into continuous series:
- *   { varX: [1, 10, ...], varY: [2. 20, ...], varZ: [3, 30,...] }
- * @param {string} csv - unparsed CSV string from a data fetch response
- * @param {Object} variables - current state.variables object
- * @return {Object} series object to be applied to state (product/site/position/month/downloadPkg)
- */
-
-
-var parseSeriesData = function parseSeriesData(csv, variables) {
-  var series = {};
-  var fields = [];
-  var rows = csv.split('\n');
-
-  if (!rows.length) {
-    return series;
-  } // Our variables list is from the most recent month. Sometimes older months have variables that
-  // are no longer around. Ultimately it would be best to proactively pull and display these, but
-  // for now we ignore any series that we don't already have variable info on.
-
-
-  var skipIndexes = [];
-  rows[0].split(',').forEach(function (fieldName, idx) {
-    if (!variables[fieldName]) {
-      skipIndexes.push(idx);
-      return;
-    }
-
-    var dataType = variables[fieldName].dataType;
-    var field = {
-      fieldName: fieldName,
-      setType: DATA_TYPE_SETTERS[dataType] ? DATA_TYPE_SETTERS[dataType] : function (v) {
-        return v;
-      }
-    };
-    fields.push(field);
-    series[fieldName] = {
-      data: [],
-      range: [null, null],
-      sum: 0,
-      count: 0,
-      variance: 0
-    };
-  });
-  rows.slice(1).forEach(function (row) {
-    if (!row.length) {
-      return;
-    }
-
-    var values = row.split(',');
-    values.filter(function (value, idx) {
-      return !skipIndexes.includes(idx);
-    }).forEach(function (value, idx) {
-      var typedValue = fields[idx].setType(value);
-      series[fields[idx].fieldName].data.push(typedValue); // Don't bother updating the range for non-numerical series and quality flags
-
-      if (typeof typedValue === 'number' && !/QF$/.test(fields[idx].fieldName)) {
-        series[fields[idx].fieldName].range = getUpdatedValueRange(series[fields[idx].fieldName].range, typedValue);
-        series[fields[idx].fieldName].sum += typedValue;
-        series[fields[idx].fieldName].count += 1;
-      }
-    });
-  }); // Loop across all numeric non-quality-flag series again to calculate series variance
-
-  Object.keys(series).filter(function (fieldName) {
-    return !/QF$/.test(fieldName) && series[fieldName].count > 0;
-  }).forEach(function (fieldName) {
-    var dataType = variables[fieldName].dataType;
-    var setType = DATA_TYPE_SETTERS[dataType] ? DATA_TYPE_SETTERS[dataType] : function (v) {
-      return v;
-    };
-
-    if (!['real', 'signed integer', 'unsigned integer'].includes(dataType)) {
-      return;
-    }
-
-    var mean = series[fieldName].sum / series[fieldName].count;
-    var sumOfSquares = 0;
-    series[fieldName].data.forEach(function (value) {
-      if (value === null) {
-        return;
-      }
-
-      var typedValue = setType(value);
-      sumOfSquares += Math.pow(typedValue - mean, 2);
-    });
-    series[fieldName].variance = sumOfSquares / series[fieldName].count;
-  });
-  return series;
-};
-/**
  * Build an updated state.selection object to fill in sane defaults from current state.
  * The goal is to have the selection always be valid and as complete as possible. For example,
  * as soon as the first variables request has completed and been parsed into state, select a
@@ -1182,7 +1072,7 @@ var reducer = function reducer(state, action) {
 
     case 'fetchDataFileSucceeded':
       newState.product.sites[action.siteCode].positions[action.position].data[action.month][action.downloadPkg][action.timeStep].status = FETCH_STATUS.SUCCESS;
-      newState.product.sites[action.siteCode].positions[action.position].data[action.month][action.downloadPkg][action.timeStep].series = parseSeriesData(action.csv, newState.variables);
+      newState.product.sites[action.siteCode].positions[action.position].data[action.month][action.downloadPkg][action.timeStep].series = action.series;
       return newState;
     // Core Selection Actions
 
@@ -1613,11 +1503,17 @@ var Provider = function Provider(props) {
             dataActions.push(actionProps); // Add a file fetch observable to the main list
 
             dataFetches.push(fetchCSV(url).pipe((0, _operators.map)(function (response) {
-              dispatch(_extends({
-                type: 'fetchDataFileSucceeded',
-                csv: response.response
-              }, actionProps));
-              return (0, _rxjs.of)(true);
+              return response.response;
+            }), (0, _operators.switchMap)(function (csv) {
+              return (0, _parseTimeSeriesData.default)({
+                csv: csv,
+                variables: state.variables
+              }).then(function (series) {
+                dispatch(_extends({
+                  type: 'fetchDataFileSucceeded',
+                  series: series
+                }, actionProps));
+              });
             }), (0, _operators.catchError)(function (error) {
               dispatch(_extends({
                 type: 'fetchDataFileFailed',
@@ -1769,8 +1665,7 @@ Provider.defaultProps = {
 var TimeSeriesViewerContext = {
   Provider: Provider,
   useTimeSeriesViewerState: useTimeSeriesViewerState,
-  TimeSeriesViewerPropTypes: TimeSeriesViewerPropTypes,
-  getUpdatedValueRange: getUpdatedValueRange
+  TimeSeriesViewerPropTypes: TimeSeriesViewerPropTypes
 };
 var _default = TimeSeriesViewerContext;
 exports.default = _default;
