+++ lib/components/TimeSeriesViewer/TimeSeriesViewerSites.js
@@ -249,7 +249,7 @@ var useStyles = (0, _styles.makeStyles)(function (theme) {
       display: 'flex',
       alignItems: 'center',
       justifyContent: 'flex-start',
-      padding: theme.spacing(0.5, 2, 0.5, 0.5),
+      padding: theme.spacing(1, 2, 1, 0.5),
       backgroundColor: theme.palette.grey[100],
       marginTop: theme.spacing(1.5),
       marginRight: theme.spacing(3)
@@ -452,8 +452,10 @@ function PositionHistoryButton(props) {
   }, "z"), /*#__PURE__*/_react.default.createElement(_TableCell.default, {
     align: "right"
   }, "Elevation"))), /*#__PURE__*/_react.default.createElement(_TableBody.default, null, history.map(function (row, idx) {
-    var start = row.start,
-        rawEnd = row.end,
+    var _row$start = row.start,
+        start = _row$start === void 0 ? '' : _row$start,
+        _row$end = row.end,
+        rawEnd = _row$end === void 0 ? '' : _row$end,
         xOffset = row.xOffset,
         yOffset = row.yOffset,
         zOffset = row.zOffset,
@@ -505,13 +507,13 @@ PositionHistoryButton.propTypes = {
     azimuth: _propTypes.default.string.isRequired,
     pitch: _propTypes.default.string.isRequired,
     roll: _propTypes.default.string.isRequired,
-    start: _propTypes.default.string.isRequired,
-    end: _propTypes.default.string.isRequired,
+    start: _propTypes.default.string,
+    end: _propTypes.default.string,
     xOffset: _propTypes.default.string.isRequired,
     yOffset: _propTypes.default.string.isRequired,
     zOffset: _propTypes.default.string.isRequired,
-    referenceStart: _propTypes.default.string.isRequired,
-    referenceEnd: _propTypes.default.string.isRequired,
+    referenceStart: _propTypes.default.string,
+    referenceEnd: _propTypes.default.string,
     referenceLatitude: _propTypes.default.string.isRequired,
     referenceLongitude: _propTypes.default.string.isRequired,
     referenceElevation: _propTypes.default.string.isRequired
@@ -1009,7 +1011,8 @@ function SelectedSite(props) {
       });
     },
     style: {
-      minWidth: _Theme.default.spacing(13)
+      minWidth: _Theme.default.spacing(13),
+      whiteSpace: 'nowrap'
     },
     disabled: disabled,
     startIcon: /*#__PURE__*/_react.default.createElement(_Clear.default, null)
